# Reduce overhead of `Waveform`