--- a/osu.Framework/Audio/Track/Waveform.cs
+++ b/osu.Framework/Audio/Track/Waveform.cs
@@ -29,7 +29,7 @@ namespace osu.Framework.Audio.Track
         /// <summary>
         /// The data stream is iteratively decoded to provide this many points per iteration so as to not exceed BASS's internal buffer size.
         /// </summary>
-        private const int points_per_iteration = 100000;
+        private const int points_per_iteration = 1000;
 
         /// <summary>
         /// FFT1024 gives ~40hz accuracy.