--- a/osu.Framework/Graphics/Audio/WaveformGraph.cs
+++ b/osu.Framework/Graphics/Audio/WaveformGraph.cs
@@ -380,10 +380,10 @@ namespace osu.Framework.Graphics.Audio
                         {
                             float height = drawSize.Y / 2;
                             quadToDraw = new Quad(
-                                new Vector2(leftX, height - points[i].Amplitude[0] * height),
-                                new Vector2(rightX, height - points[i + 1].Amplitude[0] * height),
-                                new Vector2(leftX, height + points[i].Amplitude[1] * height),
-                                new Vector2(rightX, height + points[i + 1].Amplitude[1] * height)
+                                new Vector2(leftX, height - points[i].AmplitudeLeft * height),
+                                new Vector2(rightX, height - points[i + 1].AmplitudeLeft * height),
+                                new Vector2(leftX, height + points[i].AmplitudeRight * height),
+                                new Vector2(rightX, height + points[i + 1].AmplitudeRight * height)
                             );
                             break;
                         }
@@ -391,8 +391,8 @@ namespace osu.Framework.Graphics.Audio
                         case 1:
                         {
                             quadToDraw = new Quad(
-                                new Vector2(leftX, drawSize.Y - points[i].Amplitude[0] * drawSize.Y),
-                                new Vector2(rightX, drawSize.Y - points[i + 1].Amplitude[0] * drawSize.Y),
+                                new Vector2(leftX, drawSize.Y - points[i].AmplitudeLeft * drawSize.Y),
+                                new Vector2(rightX, drawSize.Y - points[i + 1].AmplitudeLeft * drawSize.Y),
                                 new Vector2(leftX, drawSize.Y),
                                 new Vector2(rightX, drawSize.Y)
                             );