Noticed that waveform points are up in the top 5 memory usages in editor (*game*-wide) so did a bit of refactoring. There's likely much more that can be improved here, but this was the low-hanging pass.

Before:

![Parallels Desktop 2022-09-15 at 05 10 27](https://user-images.githubusercontent.com/191335/190319635-8d97a37d-a2de-485c-907b-4202c577cff4.png)

![Parallels Desktop 2022-09-15 at 05 11 14](https://user-images.githubusercontent.com/191335/190319745-b2bbb8c5-01e1-4cab-bb9f-e2e4186699d1.png)

After:

![Parallels Desktop 2022-09-15 at 05 07 45](https://user-images.githubusercontent.com/191335/190319393-dc443d7a-8d53-4bf1-95af-da20a875c42d.png)

Note that the counts here are a bit off (one more retained array in the "after" run) so savings are probably a touch better than displayed. Also the main goal was to switch away from `Point` being a `class`.

I also looked into reducing overall allocations during construction and resampling. The main fix for this was to reduce the `points_per_iteration` to a more sane number (it was previously allocating a very large buffer with no real benefit performance-wise). And then using `ArrayPool` for it (now that the size is beneath the `ArrayPool` allowance).

These results are probably more telling:

Before:

|       Method |  size |      Mean |     Error |    StdDev |    Gen 0 |    Gen 1 |    Gen 2 | Allocated |
|------------- |------ |----------:|----------:|----------:|---------:|---------:|---------:|----------:|
|   TestCreate |     ? | 99.921 ms | 1.2161 ms | 1.1375 ms | 800.0000 | 600.0000 | 400.0000 | 23,815 KB |
| TestResample |  1024 |  3.468 ms | 0.0175 ms | 0.0164 ms |   7.8125 |        - |        - |     98 KB |
| TestResample | 32768 |  6.994 ms | 0.1383 ms | 0.1480 ms | 257.8125 | 140.6250 | 109.3750 |  3,074 KB |

After:


|       Method |  size |      Mean |     Error |    StdDev |    Gen 0 |    Gen 1 |    Gen 2 | Allocated |
|------------- |------ |----------:|----------:|----------:|---------:|---------:|---------:|----------:|
|   TestCreate |     ? | 91.525 ms | 0.8581 ms | 0.8026 ms | 333.3333 | 166.6667 | 166.6667 |  3,614 KB |
| TestResample |  1024 |  2.323 ms | 0.0098 ms | 0.0082 ms |   3.9063 |        - |        - |     58 KB |
| TestResample | 32768 |  3.421 ms | 0.0214 ms | 0.0200 ms | 285.1563 | 257.8125 | 195.3125 |  1,795 KB |