@smoogipoo 2022-09-15
Clicking the "NullWaveform" button in the test scene crashes with
```
[runtime] 2022-09-15 09:42:18 [error]: System.IndexOutOfRangeException: Index was outside the bounds of the array.
[runtime] 2022-09-15 09:42:18 [error]: at osu.Framework.Graphics.Audio.WaveformGraph.WaveformDrawNode.Draw(IRenderer renderer) in /home/smgi/Repos/osu-framework/osu.Framework/Graphics/Audio/WaveformGraph.cs:line 382
```
Not sure why it's not failing on CI - pretty scary.