@peppy 2022-09-15
> Not sure why it's not failing on CI - pretty scary.

Probably this:

https://github.com/ppy/osu-framework/blob/66d72918a1fece087665c643d0cb3194790973c1/osu.Framework/Audio/Track/Waveform.cs#L87