@peppy 2022-09-15
Definitely worth the change:

|       Method |  size |      Mean |     Error |    StdDev |    Gen 0 |    Gen 1 |    Gen 2 | Allocated |
|------------- |------ |----------:|----------:|----------:|---------:|---------:|---------:|----------:|
|   TestCreate |     ? | 81.630 ms | 0.4930 ms | 0.4612 ms | 142.8571 | 142.8571 | 142.8571 |  1,317 KB |
| TestResample |  1024 |  2.057 ms | 0.0041 ms | 0.0038 ms |        - |        - |        - |     22 KB |
| TestResample | 32768 |  2.283 ms | 0.0042 ms | 0.0035 ms | 164.0625 | 164.0625 | 164.0625 |    643 KB |